refactor(core): move CV utils + clamp to nukebg-core (phase 5)
Move workers/cv/{utils,clamp}.ts and their tests into nukebg-core/src/cv (pure math, zero DOM). Add cv barrel + 'export * as cv' from core's index. Rewire 10 app cv importers to nukebg-core/cv/* sub-paths. Vite alias switched to array form with a RegExp entry so sub-path imports resolve to source (tsconfig paths already cover nukebg-core/*).

Closes #270
