+++ packages/nukebg-app/src/workers/cv/watermark-dalle.ts
@@ -1,6 +1,6 @@
 import { DALLE_WATERMARK_PARAMS } from 'nukebg-core';
 import type { WatermarkResult } from '../../types/pipeline';
-import { pixelIndex } from './utils';
+import { pixelIndex } from 'nukebg-core/cv/utils';
 
 /**
  * Detect the DALL-E 3 watermark: a multicolor bar of ~5px in the
