+++ packages/nukebg-core/src/cv/index.ts
@@ -1,0 +1,3 @@
+// CV utility barrel — re-exports pure CV helpers moved from nukebg-app
+export { RingBuffer, pixelIndex, maxChannelDiff, mean, std, median } from './utils.js';
+export { clamp255 } from './clamp.js';
